chore: Helpers class
